# Fix parsing of trailing zeros: TryParseNumber