--- a/src/libraries/System.Formats.Cbor/tests/Writer/CborWriterTests.Tag.cs
+++ b/src/libraries/System.Formats.Cbor/tests/Writer/CborWriterTests.Tag.cs
@@ -158,7 +158,7 @@ namespace System.Formats.Cbor.Tests
         [InlineData("1", "c4820001")]
         [InlineData("-1", "c4820020")]
         [InlineData("1.1", "c482200b")]
-        [InlineData("1.000", "c482221903e8")]
+        [InlineData("1.000", "c4820001")]
         [InlineData("273.15", "c48221196ab3")]
         [InlineData("79228162514264337593543950335", "c48200c24cffffffffffffffffffffffff")] // decimal.MaxValue
         [InlineData("7922816251426433759354395033.5", "c48220c24cffffffffffffffffffffffff")]