--- a/src/libraries/System.Private.CoreLib/src/System/Number.Formatting.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Number.Formatting.cs
@@ -374,7 +374,7 @@ namespace System
             }
             *dst = (byte)('\0');
 
-            number.CheckConsistency();
+            number.CheckConsistency(skipTrailingZeroCheck: true);
         }
 
         public static string FormatDouble(double value, string? format, NumberFormatInfo info)