--- a/src/libraries/System.Runtime/tests/System/DecimalTests.cs
+++ b/src/libraries/System.Runtime/tests/System/DecimalTests.cs
@@ -788,6 +788,7 @@ namespace System.Tests
             yield return new object[] { "  123  ", defaultStyle, null, 123m };
             yield return new object[] { (567.89m).ToString(), defaultStyle, null, 567.89m };
             yield return new object[] { (-567.89m).ToString(), defaultStyle, null, -567.89m };
+            yield return new object[] { "0.6666666666666666666666666666500000000000000000000000000000000000000000000000000000000000000", defaultStyle, null, 0.6666666666666666666666666666m };
 
             yield return new object[] { "79228162514264337593543950335", defaultStyle, null, 79228162514264337593543950335m };
             yield return new object[] { "-79228162514264337593543950335", defaultStyle, null, -79228162514264337593543950335m };