--- a/src/libraries/System.Runtime/tests/System/HalfTests.cs
+++ b/src/libraries/System.Runtime/tests/System/HalfTests.cs
@@ -666,6 +666,26 @@ namespace System.Tests
             yield return new object[] { (-567.89f).ToString(), defaultStyle, null, -567.89f };
             yield return new object[] { "1E23", defaultStyle, null, 1E23f };
 
+            // 2^11 + 1. Not exactly representable
+            yield return new object[] { "2049.0", defaultStyle, invariantFormat, 2048.0f };
+            yield return new object[] { "2049.0000000000000001", defaultStyle, invariantFormat, 2050.0f };
+            yield return new object[] { "5.004", defaultStyle, invariantFormat, 5.004f };
+            yield return new object[] { "5.040", defaultStyle, invariantFormat, 5.04f };
+            // input equal to half.BufferSize
+            yield return new object[] { "5.000000000000000004", defaultStyle, invariantFormat, 5.0f };
+            yield return new object[] { "5.004000000000000000", defaultStyle, invariantFormat, 5.004f };
+            // input greater than half.BufferSize
+            yield return new object[] { "5.0000000000000000004", defaultStyle, invariantFormat, 5.0f };
+            yield return new object[] { "5.0040000000000000000", defaultStyle, invariantFormat, 5.004f };
+
+            yield return new object[] { "5004.000000000000000", defaultStyle, invariantFormat, 5004.0f };
+            yield return new object[] { "50040.0", defaultStyle, invariantFormat, 50040.0f };
+            yield return new object[] { "5004", defaultStyle, invariantFormat, 5004.0f };
+            yield return new object[] { "050040", defaultStyle, invariantFormat, 50040.0f };
+            yield return new object[] { "0.000000000000000000", defaultStyle, invariantFormat, 0.0f };
+            yield return new object[] { "0.005", defaultStyle, invariantFormat, 0.005f };
+            yield return new object[] { "0.0400", defaultStyle, invariantFormat, 0.04f };
+
             yield return new object[] { (123.1f).ToString(), NumberStyles.AllowDecimalPoint, null, 123.1f };
             yield return new object[] { (1000.0f).ToString("N0"), NumberStyles.AllowThousands, null, 1000.0f };
 
@@ -991,16 +1011,16 @@ namespace System.Tests
             yield return new object[] { Half.MaxValue };
             yield return new object[] { Half.PositiveInfinity };
 
-            yield return new object[] { (UInt16BitsToHalf(0b0_00001_0000000000))}; // smallest normal
-            yield return new object[] { (UInt16BitsToHalf(0b0_00000_1111111111))}; // largest subnormal
-            yield return new object[] { (UInt16BitsToHalf(0b0_00000_1000000000))}; // middle subnormal
-            yield return new object[] { (UInt16BitsToHalf(0b0_00000_0111111111))}; // just below middle subnormal
-            yield return new object[] { (UInt16BitsToHalf(0b0_00000_0000000001))}; // smallest subnormal
-            yield return new object[] { (UInt16BitsToHalf(0b1_00000_0000000001))}; // highest negative subnormal
-            yield return new object[] { (UInt16BitsToHalf(0b1_00000_0111111111))}; // just above negative middle subnormal
-            yield return new object[] { (UInt16BitsToHalf(0b1_00000_1000000000))}; // negative middle subnormal
-            yield return new object[] { (UInt16BitsToHalf(0b1_00000_1111111111))}; // lowest negative subnormal
-            yield return new object[] { (UInt16BitsToHalf(0b1_00001_0000000000))}; // highest negative normal
+            yield return new object[] { (UInt16BitsToHalf(0b0_00001_0000000000)) }; // smallest normal
+            yield return new object[] { (UInt16BitsToHalf(0b0_00000_1111111111)) }; // largest subnormal
+            yield return new object[] { (UInt16BitsToHalf(0b0_00000_1000000000)) }; // middle subnormal
+            yield return new object[] { (UInt16BitsToHalf(0b0_00000_0111111111)) }; // just below middle subnormal
+            yield return new object[] { (UInt16BitsToHalf(0b0_00000_0000000001)) }; // smallest subnormal
+            yield return new object[] { (UInt16BitsToHalf(0b1_00000_0000000001)) }; // highest negative subnormal
+            yield return new object[] { (UInt16BitsToHalf(0b1_00000_0111111111)) }; // just above negative middle subnormal
+            yield return new object[] { (UInt16BitsToHalf(0b1_00000_1000000000)) }; // negative middle subnormal
+            yield return new object[] { (UInt16BitsToHalf(0b1_00000_1111111111)) }; // lowest negative subnormal
+            yield return new object[] { (UInt16BitsToHalf(0b1_00001_0000000000)) }; // highest negative normal
         }
 
         [Theory]