Fixes https://github.com/dotnet/runtime/issues/46827

We're currently not throwing away the trailing zeros from the fractional part of a number while parsing. This leads us to erroneously round a floating point number sometimes. This PR fixes that. I've also updated the CheckConsistency method to detect trailing zeros.