@danmosemsft 2021-01-30
Do we have perf tests.. can you get numbers?