@pgovind 2021-02-03
> Do we have perf tests.. can you get numbers?

I'm reasonably confident that CI will be green now, so I got some perf numbers:

``` Decimal
summary:
worse: 6, geomean: 1.126
total diff: 6

| Slower                                                  | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Tests.Perf_Decimal.Mod                           |      1.21 |            13.13 |            15.88 |         |
| System.Tests.Perf_Decimal.Parse(value: "123456.789")    |      1.14 |            79.49 |            90.56 | several?|
| System.Tests.Perf_Decimal.Floor                         |      1.14 |            11.65 |            13.23 | bimodal |
| System.Tests.Perf_Decimal.TryParse(value: "123456.789") |      1.11 |            80.55 |            89.61 | several?|
| System.Tests.Perf_Decimal.Divide                        |      1.11 |            66.90 |            74.11 |         |
| System.Tests.Perf_Decimal.ToString(value: 123456.789)   |      1.06 |            73.23 |            77.61 | several?|

No Faster results for the provided threshold = 0.001% and noise filter = 0.3ns.

No file given
```

``` Double
summary:
better: 4, geomean: 1.034
worse: 1, geomean: 1.027
total diff: 5

| Slower                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Tests.Perf_Double.Parse(value: "1.7976931348623157e+308") |      1.03 |           436.96 |           448.71 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Tests.Perf_Double.ToStringWithCultureInfo(value: 12345, culture: zh)      |      1.08 |           153.69 |           142.69 | bimodal |
| System.Tests.Perf_Double.Parse(value: "12345")                                   |      1.02 |            66.27 |            64.96 |         |
| System.Tests.Perf_Double.ToString(value: 12345)                                  |      1.02 |           158.42 |           155.29 | several?|
| System.Tests.Perf_Double.ToStringWithFormat(value: -1.7976931348623157E+308, for |      1.02 |           277.30 |           271.97 | several?|

No file given

```

``` Single
summary:
better: 4, geomean: 1.034
worse: 1, geomean: 1.027
total diff: 5

| Slower                                                           | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Tests.Perf_Double.Parse(value: "1.7976931348623157e+308") |      1.03 |           436.96 |           448.71 |         |

| Faster                                                                           | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Tests.Perf_Double.ToStringWithCultureInfo(value: 12345, culture: zh)      |      1.08 |           153.69 |           142.69 | bimodal |
| System.Tests.Perf_Double.Parse(value: "12345")                                   |      1.02 |            66.27 |            64.96 |         |
| System.Tests.Perf_Double.ToString(value: 12345)                                  |      1.02 |           158.42 |           155.29 | several?|
| System.Tests.Perf_Double.ToStringWithFormat(value: -1.7976931348623157E+308, for |      1.02 |           277.30 |           271.97 | several?|

No file given

```

I'm a little skeptical about these perf numbers. This PR touches only the `Parse` and `TryParse` methods. The other benchmarks also show a slow down here that I can't explain. For ex: The `ToString` and `ToStringWIthFormat` methods are untouched by this PR. All things considered, we're likely slowing down the `Parse` and `TryParse` methods a little bit here, but I'd still take the PR since we're fixing a bug here.