@pgovind 2021-02-17
I removed the trailing zero check in `CheckConsistency` now and opened https://github.com/dotnet/runtime/issues/48418 to track adding it back