@pranavkm 2021-02-19
Blazor (both .NET Core and WASM) has a couple of tests where it attempts to verify that decimal -> string conversions roundtrip without any modifications. With this change, trailing 0s for decimal values are trimmed.

* Is there a way to retain the current behavior of keeping trailing 0s?
* Have you considered treating this as breaking change?