@pranavkm 2021-02-20
Found a more egregious issue:

```
Console.WriteLine(int.Parse("3.00", NumberStyles.Number, CultureInfo.InvariantCulture));
```

Before: 3
After: 
```
Unhandled exception. System.IndexOutOfRangeException: Index was outside the bounds of the array.
   at System.Number.TryParseNumber(Char*& str, Char* strEnd, NumberStyles styles, NumberBuffer& number, NumberFormatInfo info)
   at System.Number.TryStringToNumber(ReadOnlySpan`1 value, NumberStyles styles, NumberBuffer& number, NumberFormatInfo info)
   at System.Number.TryParseInt32Number(ReadOnlySpan`1 value, NumberStyles styles, NumberFormatInfo info, Int32& result)
   at System.Number.ParseInt32(ReadOnlySpan`1 value, NumberStyles styles, NumberFormatInfo info)
   at System.Int32.Parse(String s, NumberStyles style, IFormatProvider provider)
```