@jeffhandley 2021-02-20
We will definitely need to fix this exception--that [egregious issue](https://github.com/dotnet/runtime/pull/47666#issuecomment-782471609) is release-blocking for sure. So we'll need to get either a fix for that, or we'll need to revert the change and try again in Preview 3.

The [original issue](https://github.com/dotnet/runtime/issues/46874) that this spawns from is marked as a breaking change, but we need to look at these trailing zeroes cases more closely.