@pgovind 2021-02-20
Alright, the bug occurs because of this line:

```csharp
number.DigitsCount = digEnd - numberOfTrailingZeros;
```

digEnd is 1 and numberOfTrailingZeros = 2, so we end up with -1. The correct value here should be `number.DigitsCount` = 1. I'll consider a careful fix for this locally and put a PR up.

Currently, this bug only affects `Int.Parse` calls with an input that has trailing zeros => `1.0, 1.00, 1.000, 1.0000, 1.0100 etc` are broken, but `1.01, 1.001, 1.0001, 1.01` etc are not broken.

@jeffhandley : Just wondering about time line here? After I put a PR up, I'm not sure when it'll get approved and merged :)