@pgovind 2021-02-22
> Is there a way to retain the current behavior of keeping trailing 0s?
>Have you considered treating this as breaking change?

#48608 will fix this, so you shouldn't have to do anything @pranavkm